Start moving correlation plot to own component
